requirements.txt
requests>=2.31.0
dnspython>=2.4.2
python-whois>=0.8.0
argparse>=1.4.0
beautifulsoup4>=4.12.2
aiohttp>=3.8.5
asyncio>=3.4.3
colorama>=0.4.6
tabulate>=0.9.0
jinja2>=3.1.2
certifi>=2023.7.22
urllib3>=2.0.4
tqdm>=4.66.1
python-nmap>=0.7.1
